Minor Bug fix with Canvas display size
Fixed a bug where if the Canvas Display size was greater or equal to
1000, then when displaying and using the canvas size in the drawing size
options, it included the "," which caused problems.   For now, just
removing the comma.

Also changed the javascript dialog box to allow it to close after an
error is reported (as opposed to it keep displaying the error and not
being able to continue)